+++ nucleotides/log.py
@@ -1,8 +1,12 @@
 import logging
 
 def create_logger(path):
-    logger = logging.getLogger("nucleotides")
-    fh = logging.FileHandler(path)
-    fh.setLevel(level = logging.INFO)
-    logger.addHandler(fh)
+
+    # Logging to file
+    logging.basicConfig(
+            level=logging.DEBUG,
+            filename=path,
+            format='[%(asctime)s] %(name)-12s %(levelname)-8s %(message)s',
+            filemode='w')
+    logger = logging.getLogger("nucleotides.client")
     return logger