+++ nucleotides/task/short_read_assembler.py
@@ -4,7 +4,7 @@ import nucleotides.metrics     as met
 import nucleotides.filesystem  as fs
 
 def biobox_args(app):
-    path = fs.get_input_file_path('short_read_fastq', app)
+    path = fs.get_task_path_file_without_name(app, 'inputs/short_read_fastq')
     return [{"fastq" : [
         {"id" : 0 , "value" : path, "type": "paired"}]}]
 
@@ -12,10 +12,10 @@ def output_files():
     return [['contig_fasta', [0, 'fasta', 0]]]
 
 def collect_metrics(app):
-    import json
-    path = app['path'] + "/outputs/container_runtime_metrics/metrics.json"
+    import json, gzip
+    path = fs.get_task_file_path(app, "outputs/container_runtime_metrics/metrics.json.gz")
     if os.path.isfile(path):
-        with open(path) as f:
+        with gzip.open(path) as f:
             return met.parse_runtime_metrics(json.loads(f.read()))
     else:
         return {}